+++ src/Kernel/Atomix.Kernel_H/lib/Marshal.cs
@@ -15,10 +15,10 @@ namespace Atomix.Kernel_H.Lib
 {
     internal class Marshal
     {
-        internal static unsafe void Copy(char* aDes, string aSrc, uint aLen)
+        internal static unsafe void Copy(char* aDes, string aSrc, int aLen)
         {
             uint src = Native.GetContentAddress(aSrc);
-            Memory.FastCopy((uint)aDes, src, aLen * sizeof(char));
+            Memory.FastCopy((uint)aDes, src, (uint)(aLen * sizeof(char)));
             aDes[aLen] = '\0';
         }
     }
